pyproject.toml
[project]
name = "UAP"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Ganesha Mahardika P", email = "ganeshaprasetya1@gmail.com"},
]
dependencies = ["tensorflow>=2.20.0", "scikit-learn>=1.8.0", "matplotlib>=3.10.8", "seaborn>=0.13.2", "streamlit>=1.52.2", "pillow>=12.0.0", "ipykernel>=7.1.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
